+++ data-raw/data-raw.R
@@ -1,0 +1,3 @@
+source("data-raw/fishery.R")
+source("data-raw/gerrard.R")
+source("data-raw/kokanee.R")